requirements.txt
streamlit
langchain
langchain-groq
langchain-pinecone
sentence-transformers
langchain-community
pinecone-client
langchain-core
groq